flow: Add a regression test for architectures with cin tie off connections
This regression test tests the code that supports architectures
with cin tie off connections. This is the case when the cin pin of the
cluster is connected through muxes to more than one adder in the cluster.
Allowing more than one starting point for the adder chain within the cluster.