infra: Support a BUILD_DIR value outside of source tree.
Allows setting `BUILD_DIR` to a value outside the build tree.

If `BUILD_DIR` was set to anything other than a directory inside the
source tree it would fail as previously the Makefile did `cmake ..`.

Symlink `build` directory to the `BUILD_DIR` location.

Add symlinks for the old binary paths.