vpr: Fix power related memory leak in pb graph interconnect pins
For reasons that aren't entirely clear power analysis requires knowing
the t_pb_graph_pins associated with each interconnect block described in
the architecture, which are allocated in a somewhat bizarre data
structure by VPR.

Previously they were not freed with the rest of the pb graph. If this
code were re-architected, we shouldn't have VPR loading a sub-set of the
pb graph when the main pb graph is being freed by libarchfpga.

For now we just ensure that libarchfpga frees the power/interconnect
related pb graph components correctly.