build: Bump cmake version to 3.9 & fix LTO builds
Previously we applied LTO flags only to vpr, which meant other libraries
(e.g. libvtrutil, libarchfpga) which are linked into VPR were built
without LTO flags. This is suspected to be the cause of compiler crashes
with LTO on gcc 7/8/9.

Requires cmake 3.9 to consistently set LTO flags on all parts of VTR
(with the exception of ABC).