fasm.cpp: return from visit if pb is `open`
This change avoids the net visitor to write fasm outputs of
pb_graph_nodes which do not have valid corresponding PBs and valid
parent PBs (this could suggest that the PB is actually `open`).

These PBs are skipped and the visit goes on to the next PB in the net.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>