flow: Fix non-deterministic clock domain black boxing
Previously the script used to identify clock domains (for black-boxing
each clock-domain before passing the circuit through ABC) relied on
perl hashes to store the clock domains. While this was functionally
correct it reported the clock domains in a non-deterministic order,
(since the of elements in a perl hash are not ordered).

This made the entire flow non-deterministic (for multi-clock circuits)
since ABC produced a different circuit implementation depending on the
clock domain processing order.

To fix this the script now sorts the clock domains before they are
reported, ensuring a consistent deterministic order is uesd.