vpr: Re-work error suppression to use VPR_ERROR() instead of VPR_THROW()
Historically VPR_THROW() has been used to unconditionally throw an
exception (to signal an error). Since some parts of the code rely on
this unconditional behaviour, VPR_THROWs should not be suppress-able.
(If VPR_THROW is allowed to be suppressible a number of new compiler
warnings are generated since calls to it can now return.)

Instead, this commit adds VPR_ERROR() which signals an error which (by
default) throws an exception, but which may be converted/supressed to
a warning (i.e. to signal a non-fatal error).

* VPR_THROW() should now be used for unconditional fatal errors.
* VPR_ERROR() should now be used for suppressible (non-fatal) errors.

Since the code-base did not previously make the distinction between
fatal/non-fatal errors, there are likely a number of legacy VPR_THROWs
(but not all) which can be converted to VPR_ERRORs in the future.