odin: makefile: fix parallel build
There were two issues with the way the targets are defined.

1. The .%.build target cannot depend on both clean and _init. It works
fine with a single-core build, but when executing it in parallel, the
clean target breaks the _init target. To make it work in parallel the
_init target now directly depends on the clean target.

2. The _set_<xxx> targets have to be executed before the _build_it
target - they cannot be executed at the same time. This has been
reworked to use a dynamically defined targets to avoid creating an extra
wrapper target for each of them.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>