Use FASM prefix of parent pb_type, not prefix of output pb_type.
The previous code used the clb prefix of the current block, but the
interconnect/fasm_mux logic operates on the parent pb_type of the
current block, not the current block itself.  As a result, if the output
of a fasm_mux has a prefix, it was getting lumped into the FASM mux
declaration.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>