equivalent: reorder physical/logical equivalent tiles based on pin differences
This change enables a sorting of the logical block and physical tiles
equivalent counterpart lists based on the differences in the number of pins.

This ordering allows to select the "best" corresponding candidate of a
logical/physical type when there is the need to access its data.

This also allows to not rely on the `pick_random_type` function

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>