VTR 7.0 / VPR 7.0 Release
Released: 2013-05-13

Downloaded from: http://www.eecg.utoronto.ca/vtr/vtr_release_7.0_full.tar.gz

Starting from this release, the VTR version number was bumped to align
with the VPR version number.

The existing mainline 'master' branch revision history contains the
period during which this release was made. However this release also
appears to also have other out-of-tree modifications, and so it does
not correspond directly to a commit in the mainline revision history.