vpr: Reduce RRGraph memory usage by merging node edge arrays
We now use a single array to store the edges associated with a node.
The array is sorted into sub-ranges (e.g. incoming/outgoing,
configurable/nonconfigurable) which still allows efficient iteration
through the edge subranges. We store the sizes of the various sub-ranges
to enable quickly determination of the delimiters between sub-ranges.

We also use a raw pointer for the edge array (rather than a vtr::vector)
which further saves memory.

The create_edge() routine now no longer inserts the edge into the node
data. Instead a single call is made to rebuild_node_edges() which will
walk the various RRGraph::edge_* members to precisely allocate the
relevant node_edges_ and node_num*edges_ data members.

The trade-off of this change is that the various node_*_edges() will not
work properly after an edge is added to the RR graph until
rebuild_node_edges() is called. Currently it rebuilds the node edges
from scratch, and so calls to it should be minimized (the preferred
approach is to add all edges and then call rebuild_node_edges() once).
If needed, support for incrementally rebuilding node_edges
could be added (while the use of a single array per node type in general
require insertion not at the end of the array, an O(n) operation, the number
of edges per node is typically bounded by a small constant, so it would
still be reasonably efficient); it is not currently required and so is
left as *potential* future work should the need for it arise.