vpr: Pop remaining nodes off heap to improve router debug graphics
When viewing router expansion in graphics for debugging it is helpful to
see all nodes pushed into the heap (even if they were not popped before
the router found the target).

This commit pops such nodes off an annotates their cost so they can be
viewed in the router graphics. Since this is only helpful as a debugging
feature (with some run-time cost otherwise), it is only enabled when the
--router_debug* options are enabled.