unify shebangs, make script files executable
The motivation in this change is to make it easier to run under NixOS,
which does not support absolute paths.  Note that a part of the scripts is
already using the `#!/usr/bin/env ...` shebang, which correctly calls the
executable which is present in the user's environment.  This commit makes this
behavior consistent over all scripts.

Also, the executable flags were missing for part of the scripts.