Integrate interned strings with VPR metadata, and replace node/edge unordered_map.
The unordered_map used for node/edge metadata is very expensive compared
with a flat data structure.  During reading, the data structure is
always ready for querying, but this is not required.  The new strategy
is to accumulate all the data, and only setuping the map once needed.
This saves memory (1.6 GB) and CPU time (12 seconds).

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>