Refactor some routines inside deeply nested loops to avoid reallocation.
The refactored functions:
 - get_rr_node_indices
 - label_wire_muxes
 - label_incoming_wires
all are invoked inside of nested loops, resulting in a large (~10M+)
number of repeated allocations and deallocations.  The heap may be able
to avoid fragmentation, but small code changes could result in fragmentation
where none existed before. The change is to hoist the storage container
(std::vector) outside of the loops so that once a high watermark is
reached, no further allocations will result.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>