vpr: Update block pin index to name to match new storage convention
When the logical and physical block types were seperated the storage
convention for ports was changed to be more explicit. This updates the
reverse look-up to be consistent with that format (for most
architectures the previous code was correct, but that was because most
architectures obey the same pin ordering conventions as before) which
should be more robust.

Also add some additional sanity checking to RR node debug print
routines.