vpr: Add DeviceContext::rr_segments to describe segment types in RR graph
Historically there are multiple copies of the wire segments types in the
RR graph. These exist in t_arch::Segments, and t_vpr_setup::Segments. At
the moment these are not consistent (e.g. setup_clocks appends segments
to t_vpr_setup::Segments, but other code refers to t_arch::Segments
which is out of date).

This commit ensures that the fully up-to-date segment info is stored in
device_ctx.rr_segments. In future commits the down-stream customers who
rely on this data (post create_rr_graph()) will be converted to refer to
DeviceContext::rr_segments.