vpr: Improve robustness of map lookahead at narrow channel widths
At small channel widths (e.g. with small circuits) some block pins may
be disconnected. Previously the map lookahead would return the largest
non-infinite float value to indicate this was not a wise path to
explore (since the lookahead has effectively already figured out that
there is no such path in the RR graph).

However when the lookahead was used for delay estimation to calculate
initial criticalities this value would overflow to infinity.

With this change we return a large (but not the largest) value in this
case which avoids issues in the timing analyzer.

This would only occur when the actual architecture (channel width) was
unroutable (since the pins were disconnected) so it shouldn't have any
impact on practical circuits. However this change ensures that the
router makes this determination in the standard way.