Fix compilation of vtr_vector for specializations without data().
The using storage::data method forwarding only works if there is such
a method defined in the base class, which might not be true for all
specializations (I ran into this with a vector<bool> specialization).

This can be fixed by forwarding the method in a traditional way with
an inline implementation; then, the compiler will defer compiling it
only when actually needed.

Signed-off-by: Henner Zeller <h.zeller@acm.org>