vpr: Fix inconsistent pin criticality during incremental updates
Previously, for primitive sequential pins, which may map to multiple
timing graph nodes, we set the pin criticality of the criticality of
the 'last' timing graph node to be processed (which was arbitrary).

To ensure consistent results with incremental analysis (where not all
nodes are processed) we now only use the 'external' facing timing graph
node. This ensures a consistent result, and also matches the historical
behaviour (before incremental updates were introduced).