vpr: Ensure modified timing graph edges are invalidated in router
This ensures that when run with --timing_update_type 'incremental' the
router produces the same results as when run with 'full'.

It is still future work to optimize the router to exploit incremental
STA capabilities (i.e. various parts of the router still iterate over
all netlist connections updating things).