Regression_test: Created benchmarks for each gate (ie and, xor, etc) that test the width capabilities and different methods of calling said gates. Some .vh tests call a .v test that is standardized and rely (#1335)
on the width to be define as well as the operation and possibly if it's binary or unary. There is a wire test (1 bit wide), an index test and then 3 seperate width tests: wide, int_wide, ultra_wide, which
are 3 bit wide, 32 bit wide and 256 bit wide respectively. These seperate width tests are called in two different manners. The range catagory calls on a gate with the range inside (ex not([2:0]out,[2:0]in))
while the replicate catagory calls on the gate with an outside array (ex not ar[2:0](out,in))

Co-authored-by: emacdo12 <emacdo12@unb.ca>