flow: Ensure valgrind output appears in tool logs
Previously it was re-directed to a different file, which prevented it
from appearing in failure logs printed by run_vtr_flow (making the
source of the failure rather mysterious).

Also change the valgrind error exit code to 22, which doesn't overlap
with normal exit codes (e.g. 0/1/2 as used by VPR).