Updates to lookahead map.
  - New sampling method for lookahead map.
  - New spiral hole filling algorithm
  - Use a lightweight version of PQ_Entry (PQ_Entry_Lite)
  - Use maps in run_dijkstra()
  - Move context of router_lookahead_map_utils inside of util
    namespace.
  - Parallelization of lookahead generation

lookahead: add new extended map lookahead

This lookahead generates a map with a more complex and complete way of
sampling and expanding the nodes to reach all destinations.

The dijkstra expansion is parallelized to overcome the higher number of
samples needed.

It also adds a penalty cost factor based on distance between the origin
and destination nodes.

extended_lookahead: cleaned code and reduced code duplication

Signed-off-by: Dusty DeWeese <dustin.deweese@gmail.com>
Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>
Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>