Fixing documentation errors found within the Architecture Reference section
The first error was found under the "row" tag description. Originally,
one of the optional attributes was "icry", this was changed to "incrx".
Since horizontal increments are in the x-axis.

The second error was found under the "region" tag description.
Originally, the description for the optional attribute "incry" was "An
expression specifying the horizontal increment between block
instantiations within the region.", horizontal was changed to vertical.

The third error was found under the "equivalent_sites" tag description.
Originally the note under the "direct" tag said "This tag is need only
if the pin_mapping of the site is defined as custom", need was changed
to needed.