Add definition of the funciton generate_opname_stratix10
This function takes as input a pointer to the current primitive. It then
finds the parameters indicating the operation mode and whether the in/out
ports are registered. Based on the value of these parameters, the
function generatates a mode-hash string to be used as the primitive name
in the output blif file.