Look for an empty location in find_compatible_compressed_loc_in_range().
find_compatible_compressed_loc_in_range() searches for a compatible location for a specific block type in a range around a given location. The returned location is not necessarily empty. This does not cause any problems for SA because two blocks can be easily swapped. However, in initial placement, we want to place blocks in empty locations. If this function returns a location which has already been occupied, initial placement resorts to random placement. The modified function get a new argument (check_empty) to indicate that the returned location must be empty.