[Route][Types] Removed prev_node from t_rr_node_route_inf
Removed the prev_node member of t_rr_node_route_inf since the
information can easily be found by getting the source of the previous
edge. This will reduce the size of the data that needs to be collected
during routing. This may also make the route function a bit faster since
the previous node does not need to be stored during the wave expansion;
however, this may make reconstructing the path slower.

Also updated the comment in the t_rr_node_route_inf struct regarding the
prev_edge member, since it was no longer accurate. The edge ID used to
be an index into the edges of the previous node; however, that has since
changed to a globally unique ID within the RR Graph.