[CAPNPROTO][CMake] Fixed Build When CAPNPROTO is Disabled
There were many build errors that arose when CAPNPROTO was disabled. It
looks like there was a dependency in libarchfpga on libvtrcapnproto even
though the library is disabled if the user passes a flag to disable it.
Removed this dependency when the flag is disabled.

VPR also had a similar issue. This appears to have been handled in the
past, but has since regressed. We should consider adding this build to
CI to prevent further regressions (if this feature is needed).