[Route][Heap] Removed prev_node From t_heap
A previous commit removed prev_node from t_rr_node_route_inf. This was
because the prev_node can be uniquely identified by the prev_edge (but
not the other way around).

The same methedology can be applied to the t_heap type. This was only
being set, but it was never being used (with the exception of one debug
message). Therefore this is only good for performance and memory
overhead.