[ParallelRouter] Simplified the Heap Popping
Made the heap popping mechanism and how the result is returned better
match the ability of the MultiQueue to make integration easier.

In the future it would be nice if the MultiQueue could be extended.

Also revisited the pruning functions since removing the exit condition
lead to huge degredations in performance. I am not sure if my changes
removed determinism, but the performance is now on-par for smaller
circuits.

Running the strong test, we may have some infinite loops. Need to
verify.