[ParallelRouter] Fixed Bug With Algorithm
Bitcoin miner, as well as other hard regression tests were failing. It
seems like when circuits are on the edge of routability they cause an
issue for the deterministic version for some reason. Was getting strange
allocation issues.

Turned off determinstic parallel routing for now until the parallel heap
is implemented.