[ParallelRouter] Added Multi-threading for the Router Algorithm
The components for multi-threading are added, including the thread pool,
the spin locks, and the thread-safe cost updating function.

The setting of (#T=1, #Q=2) has passed the VPR strong regression tests.
But the setting with more threads tends to run for a very long time. We
need to further investigate the multi-threading implementation.