[Yosys][Warnings] Fixed Fatal Not Git Repo Warning
Fixed a warning where "fatal: not a git repository: './/.git' was being
printed during builds of VPR.

This is caused by Yosys being included as a subtree instead of a git
submodule. The Makefile in Yosys is trying to get the current version of
git for Yosys but it cannot do it since it is not a git repo.

Just commented out the line causing the warning for now.