[ParallelRouter] Fixed Bugs and Updated the Multi-Threading Strategy
The behavior of the weird bug in the previous parallel routing algorithm
is that the single-thread setting works fine, while the multi-thread
settings stall and do not make forward progress. This is because we used
a while-loop of MQ_IO::tryPop() to clear the priority queues in the main
thread. And the tryPop function will spin forever in the case where (1)
queues are empty and (2) the number of threads spinning in tryPop (which
is exactly one) is less than the total number of threads (which is more
than one in multi-thread settings) used in MQ. The bug is fixed by
simply removing the MQ clearing function since the queues are guaranteed
to be drained in the MQ-based parallel routing algorithm.

The second part is that we updated the multi-threading strategy. We use
the main thread and multiple long-lived sub-threads as the workers for
parallel A* search. We also use two barriers to synchronize the workers
so that the router can efficiently support multiple times of single
connection routings during its lifecycle (instead of spawning worker
sub-threads for each connection route).