[RR_GRAPH] Fixed Bug in LLVM17 Build
The std::sort function requires operands to be "pointer-like", this
requires that de-referencing a const pointer should be allowed.

This was not an issue in the past, but in LLVM17 this property is
required. The edge_sort_iterator did not follow this property.

Fixing this by naively allowing this property led to a 10% performance
degredation due to requiring pointers to be allocated and deallocated.

To fix this for now, made the sort use std::stable_sort instead. The
performance degredation is quite small and may be worth it.