Fixed Make Being Called With Multiple Threads
CMake was calling another Make command, but was explicitly passing the
number of threads to use. Make files should not pass the number of
threads to use since it may cause more threads to be consumed than what
the user intended.

For example, if the user used `make -j10`, the parent make would use 10
threads, and the child make would also use 10 threads; but this is not
what the user wanted. I think CMake recognized this and was suppressing
this behaviour anyways. I believe that CMake will pass the number of
threads for the child make to use anyways, but in a way that would
prevent the above issue.

Removed the ability to set the number of threads for Yosys builds to
handle the warning.