[ParallelRouter] Determinism Improved
Found a small oversight with how the equality condition was being
pruned. This was caused when the backwards cost was equal and the total
cost was not, which is quite likely.

Fixing this change allowed the astar_fac to increase up to 0.6! I ran it
at 0,7 and sometimes it got the same answer as 0.6 but sometimes it
didnt. This implies that it is just on the edge of being deterministic.

Ran 0.6 4 times and always got the same solution on MCNC. Still need to
test on other architectures and circuits.