[ParallelRouter] Cleaned Up Deterministic Pruning Code
Tried to cover all bases for the determinstic pruning code.

Added a hard-coded factor used to get determinism. A factor of 0.4 seems
to get determinism for the MCNC benchmarks, does not work for Koios.
Koios gets determinstic CPD, but not wirelength for some reason.

Added more comments to make code clearer.