[CI] Made the Regression Tests Error on Warning
To prevent warnings from showing up in the build in the future, made the
CI error on any warnings during the regression tests.

It looks like VPR originally had the ability to error on warning;
however, it was specific to only VPR. This CMAKE method will now work
for all executables compiled during the tests. May want to consider
removing it in the future.