[CI] Added GCC-13 as a Compatability Test
Ubuntu 24.04 uses GCC-13 as its default GCC version. The CI should
ensure that VTR builds for this version of GCC.

Also added the version of Clang that would come with Ubuntu 24.04

Versions of tools were found from here:
https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2404-Readme.md

Added as its own workflow which is manually dispatched since the Noble
runners are still in beta.