[Capnproto] Fixed CMake Compatibility Error
On CI, A CMake deprecation error occurs since CMake version < 3.5 will
be removed in a future version of CMake.

On the dev branch of CapNProto, they fixed this issue by setting the
minimum CMake version to 3.16:
https://github.com/capnproto/capnproto/commit/2354ae2c75bb761f44444fa5064ec49a3d18a7da

Replicated these changes locally. Hopefully the next time we upgrade
CapNProto this will be in the changes.