[CI] Changed the Workflow Triggers
Changed the workflow triggers for the test and container workflows such
that they only run when there is a push to master.

Also added concurrency such that if a CI was called on a particular PR
or branch that is already running the CI, it will cancel the old run.
This will prevent useless runs of the CI.