[CI] Cleaned Up How Artifacts Are Saved
When artifacts were being saved for the regression tests, it was saving
all of the benchmarks as blif files. This is incredibly wasteful (72 MB
per regression test or around 720 MB per CI run). These tests will no
longer save .blif files as artifacts anymore. If anyone needed these
files they could generate them themselves or can add a condition to add
the exact .blif files they may need.

The CI was also always generating artifacts (at least for the nightly
tests) even when the run was cancelled. Made it so the artifacts are
only saved if the run is not cancelled. This would mean an artifact is
generated if the run succeeds or fails.