[CI] Prevented Running the CI If Only the Documentation Is Updated
Added a `path-ignore` filter to the GitHub Actions triggers (including
`push` and `pull_request`) to filter out any push or pull request that
only includes docs updates.

The "Test" and "Containers" workflows will be skipped if a push or a PR
only contains changes to the documentation, such as the developer guide
and README, so that we can save CI resources for other uses.