[Route] Removed Unreachable Code
In an old version of the router, a catch was made to exit the router
with no solution and not to retry at max bounding box; however, the
function was rewritten to explicitly do this. The old catch statement
was left in the code and was unreachable.

Removed the unreachable code.