[Warnings] Converted an int into a size_t
This code previously used an int to calculate the size of an allocation,
which was done by shifting a bit; however, this is technically
ill-formed since this could generate a negative value. This was causing
a warning to appear in the most recent version of GCC. Although this is
technically ok, its best that this properly uses size_t instead.