[AP] Created AP Netlist and Fixed Bug in B2B
Instead of iterating over the AtomNetlist, created our own netlist
object which is a vector of vectors. This can be made much more
efficient by using the Netlist class as a base class; however, doing
this now to make is easier.

In the process of implementing this feature, fixed a bug in the B2B
formulation which was causing it not to converge.