[VTR][LOG] Improved Unused Arg Warning Suppression
Previously, VTR's logging used `sizeof` static casted to void in order
to suppress the unused argument warning when we want logs to be a NOP.

The `sizeof` operator may have some strange affects, but since the code
is NOP anyways thats not a huge deal. The reason for this change is that
intellisence that use clangd flag this as a potential issue causing
warnings.

This also exposes a potential issue where a LOG message is using a
function that may have memory effects, so it may not be cleaned up by
the compiler properly. The unused variables should be handled
explicitly.

Also moved static variables from the header file of vtr_log to the
implementation. No functional change, but just prevents these variables
from being present and accessible in every file that uses vtr_log (which
is alot).