[CI] Separated Out NightlyTests
Separated out the NightlyTests from the regular tests. This separation
makes sense since the NightlyTests take so much longer to run and are
run on different machines. This makes it easier to configure when these
tests are run in the future (especially if we need to change the testing
machines and do not want these tests to run as often).

Also made them run nightly instead of on push to master or PRs since
they are currently not working. Once they start working again, we can
re-enable this feature if we want.