[Warnings] Fixed Const Reference
When a class returns a reference to a dereferenced pointer, if the
pointer does not point to a const object, returning that pointer as a
const has no effect. This was causing warnings.

To fix this, we want a refernece to a const object not a const reference
to an object (silly, I know). So we need to make this more clear in the
class.