[Warnings] Removed -lz Flag from Compilation
The -lz flag is used to link with the ZLIB library, but the library was
already being linked; so there was no reason to add the flag. This was
causing an insane number of warnings in the Clang build. If this is
added back for any reason, this should be checked.