[LOGGING] Fixed LUT Sizing Issue
The size of the LUTs were being calculated incorrectly by using the width
of the port instead of the actual number of pins in the port.

The parsing scripts for the golden results were also incorrect since
they were using the number of 6-LUTs as the total number of LUTs.