[Prepacking] Re-Designed Prepacker API
The original Prepacker API created a lot of global variables which were
confusing to work with and hard to maintain.

Created a Prepacker class which maintains the state of the Prepacker and
provides methods to access prepacking information.

I also tried to remove the accesses to the global scope from within the
prepacker to isolate it from the rest of VTR. This is in an attempt to
make it easier to work with.