[Router] Removed the R_upstream field in `t_rr_node_route_inf`
The `R_upstream` field added to the `t_rr_node_route_inf` structure for
convenience in commit b11b2 is removed and kept inside the connection
router class instead.

This allows the size of each `t_rr_node_route_inf` to be as minimal as
possible, which tends to be more cache-friendly when accessing the large
RR node route info data, leading to higher performance.