[Router] Refactored `t_heap` to `RTExploredNode`
The `t_heap` data structure is no longer needed as the "heap node
structure" due to the algorithmic changes in the connection router in
commit b11b2eb.

However, `t_heap` is used in many places as a holder for the return
values of the route routines.

This commit refactored the `t_heap`, renamed it to `RTExploredNode`, and
combined it with the other similar node info data structure, `node_t`,
in the connection router to fit the changes since commit b11b2eb.