[AP] Created The APNetlist Class
The APNetlist class holds the blocks and nets used during the AP flow.
In the AP context, a block is a collection of primitives which would
want to be placed in the same location (i.e. they want to move
together). Currently, each block represents a pack molecule (created by
the Prepacker). The nets are logical connections between blocks. The
APNetlist will not contain any nets which are not needed for AP.

The APNetlist is tested using a Unit Test which tests its basic
functionality (features that differ from the base Netlist class).

This is the first of many commits which will bring Analytical Placement
into VPR.