[AP] Added Partial Placement Object
Added the partial placement object. This object is used during the
Analytical Placement flow to store the positions of the APBlocks. This
placement is not gaurenteed to be legal; however, it does gaurentee that
the blocks are placed on the device and the fixed blocks are placed in
valid locations.