[TestScript] Enhanced the Robustness of the VTR Flow Log Parser Script
Added affixes to the regex pattern string, i.e., `$.*(my pattern).*$`
to ensure that the searched pattern is not strictly required to start
from the beginning of the string or line; it only needs to be part of
the line.

This makes the definition of the pattern rules easier and makes the VTR
flow log parser more robust.