[Router] Fixed the Logging Issues and Reverted the Changes for RCV Tests
Fixed the output of the SSSP time of the connection router, which caused
VTR flow log parser issues and made CI mad.

Reverted the changes for RCV tests, which are not working and appear to
be too simple to be useful. It needs more time and effort to make a
comprehensive set of regression tests for the RCV technique. We tend to
postpone this task since it is not the main purpose of this PR.