[AP] Full Legalizer
The Full Legalizer in the AP flow takes a partial placement (of any
legality) and returns a fully legal placement. A legal placement is a
set of legal clusterers of atoms and legal places for those clusters to
be placed.

To fully test the AP flow, made a temporary Global Placer which will
just place the APBlocks at the center of the device. This will be fixed
later.

Also includes basic tests to ensure the AP flow does not regress.