[AP] Analytical Solver
This introduces the Analytical Solver class to the AP flow. This is an
integral part of the Global Placement stage and what gives Analytical
Placement its name.

This PR introduces the QP_HYBRID analytical solver which uses a hybrid
Clique and Star net model to optimize the quadratic HPWL objective.

The code is designed to allow for other analytical solvers to be
implemented and interchanged without issue. A B2B solver will be coming
in a future PR.