[CI] Improved How The CI Handles Num Cores Used
Found that the CI was always building and running using only 2 cores on
the regular tests. However, the CI has 4 cores that it can use. The
number of cores can change based on some factors.

Added the ability of the tests to check how many cores they have
available, and use all the cores to build and run VTR tests.