[AP] Global Placer
Created the Global Placer base class which will create a partial
placement based on the netlist and the architecture. This attempts to
find a "globally" good placement, without considering all of the complex
constraints of the FPGA architecture (without clustering).

Implemented a SimPL-based Global Placer which maintains an upper and
lower bound solution which slowly approach each other over several
iterations. The lower-bound is the analytically solved solution which
tries to optimize the placement (hinted by the upper bound solution).
The upper-bound solution is the lower-bound solution which has been
partially legalized.