[Place] Independent Placement Verification
Created a method that would independently verify the placement in the
VPR flow. If a placement passes this verification, it is assumed that it
can be used in routing without issue.

By design, this method does not use any global variables (everything
needs to be passed in) and recomputes everything that it does not
assume. This allows it to be independent of the placement flow being
used, so this method can be used in the AP flow as well.