[Clustering] Independent Clustering Verification
Created a method that would independently verify the clustering in the
VPR flow. If a clustering passes this verification, it is assumed that
it can be used in placement and routing without issue.

By design, this method does not use any global variables (everything
needs to be passed in) and it recomputes everything that it does not
assume. This allows it to be independent of the packing flow, so this
method can also be used in the AP flow.