[STA] Fixed Incremental Slack Timing Analysis Verify Code
In debug mode, the STA checks if the incremental STA matches the STA
computed from scratch. In this check it checked for if the critical
nodes matched; however, the nodes returned are not stably deterministic
(it does not break ties in a deterministic way). After discussing with
Vaughn, it is safe to just remove the nodes check since it would not
change the result of the analysis.

closes #2754