[Router] Fixed the Bug Causing Tons of Nightly Test Failures
The bug was introduced by me in commit 2966d66; I merged two adjacent
if-bodies with conditions that appeared the same, but were in fact not.

A simplified example is:
`if (A == 0) { ... A might be modified ... } if (A == 0) { return ; }`

Merging the if-bodies would be problematic if A is modified in the first
if-body and is no longer 0.

This mistake caused tons of nightly test failures, as mentioned in the
comments in PR#2720. So glad that the nightly tests CAUGHT this bug,
which was not detected by the VTR strong regression tests.