[AP] Testing Infrastructure
Added changes required to test the AP Flow on the MCNC and VTR
benchmarks.

This creates fixed device sizes for each benchmark which can be used to
fairly compare the AP Flow to the default VTR flow.

Added the currently passing MCNC and VTR benchmarks to the strong tests
to lock in the AP flow. It is a bit unstable right now until the full
legalizer is put in. May need to disable a few of the tests.