[Placement] Resolved Inconsistency with Placing at Root Tile Locs
Found an inconsistency with how the placement of cluster blocks into
large tiles was handled in VPR. During placement, clusters are only
placed at the rool tile location of the grid; however, specific parts of
the later stages require that clusters are placed at all tile locations
of large tiles. This is inconsistent and makes it confusing to know
where clusters should be placed.

Enforced that clusters must be placed at the root tile location and
never placed in other tile locations in the large tiles. This required
fixing an oversite in the flat router which was relying on the fact that
clusters will not be placed at the root tile location.