[CMake] Upgrade Minimum Version to 3.10
The CI currently always runs on the latest version of CMake. I am not
sure why this is, its not the best practice, but CMake is a fairly
stable application so it should be fine.

CMake is deprecating version 3.9, and a few CMake files require the
minimum version to be 3.9 or larger. Changed these files to be 3.10 or
later.