[Packer] Created GreedyClusterer Class
Began encapsulating the different parts of the packer into classes. This
will help organize the packer better, which will make it easier to
modify in the future.

My plan is to clean up the different parts of the packer so I can add
flat placement information into the gain calculation so it can be
integrated into the AP flow.