[Packer] Greedy Clustering Candidate Selector Class
Outlined all of the logic related to selecting a candidate molecule to
pack into a given cluster into its own class. This class will maintain
the information required for computing gains and selecting the highest
gain molecule to pack.

By separating out this logic, it makes it easier to modify and
contribute to this code since it is now obvious where the code starts
and where it ends.

I have also went through and tried to modernize anything I could
(for example, replacing C-style arrays for vectors). This code still
needs a lot more cleanup, but this PR will be a good stepping stone for
future refactorings.