[NDMatrix] Resolved Dangling Reference Warning
GCC13 introduced a warning for potentially dangling references. The
NDMatrix class was creating temporary objects to store the offseted
pointer and then returning refrences to these pointers. Although this is
correct, since the pointer is owned by the base NDMatrix object, it was
confusing the compiler which is not something you want to do in general.

Changed how the NDMatrix proxy class works by storing the offset into
the pointer directly and passing a reference to the base pointer
instead. This makes it clear to the compiler that this pointer belongs
to another class and we are returning a reference to a portion of that
memory, which is allowed.